requirements.txt
boto3
requests
PyNaCl